+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Release 5.7.6 (unreleased)
  * Credentials: Fix displaying age verification credential
+ * Support URL scheme `av` for age verification
 
 # Release 5.7.5
  * Update to VC-K 5.10.1